+++ src/main/java/kitchenpos/eatinorders/dto/OrderTableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.eatinorders.dto;
 
-import kitchenpos.eatinorders.todo.domain.ordertable.OrderTable;
+import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
 
 import java.util.UUID;
