+++ src/main/java/kitchenpos/eatinorders/todo/domain/ordertables/OrderClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.todo.domain.ordertable;
+package kitchenpos.eatinorders.todo.domain.ordertables;
 
 public interface OrderClient {
     boolean containsInvalidOrderForClearOrderTable(OrderTable orderTable);
