+++ src/main/java/kitchenpos/eatinorders/todo/domain/ordertables/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.todo.domain.ordertable;
+package kitchenpos.eatinorders.todo.domain.ordertables;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
