+++ src/main/java/kitchenpos/eatinorders/ui/OrderTableRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.eatinorders.ui;
 
 import kitchenpos.eatinorders.application.OrderTableService;
 import kitchenpos.eatinorders.dto.OrderTableResponse;
-import kitchenpos.eatinorders.todo.domain.ordertable.NumberOfGuests;
-import kitchenpos.eatinorders.todo.domain.ordertable.OrderTableName;
+import kitchenpos.eatinorders.todo.domain.ordertables.NumberOfGuests;
+import kitchenpos.eatinorders.todo.domain.ordertables.OrderTableName;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
