+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,10 +3,10 @@ package kitchenpos;
 import kitchenpos.eatinorders.domain.Order;
 import kitchenpos.eatinorders.domain.OrderLineItem;
 import kitchenpos.eatinorders.domain.OrderStatus;
-import kitchenpos.eatinorders.todo.domain.ordertable.NumberOfGuests;
-import kitchenpos.eatinorders.todo.domain.ordertable.OrderTable;
+import kitchenpos.eatinorders.todo.domain.ordertables.NumberOfGuests;
+import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
 import kitchenpos.eatinorders.domain.OrderType;
-import kitchenpos.eatinorders.todo.domain.ordertable.OrderTableName;
+import kitchenpos.eatinorders.todo.domain.ordertables.OrderTableName;
 import kitchenpos.menus.tobe.domain.menu.Menu;
 import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import kitchenpos.menus.tobe.domain.menu.MenuName;
